Master file for Plotting analysis
This file has 2 folders, CSVDump and 23DecSOSData, which have inner folders. 
Each of them have sub folders with proper names and each of the corresponding images and the CSV files, along with the ipynb files which are to be opened in the jupyter notebook. 
Download the master file and import in the main SoS folder on your PC to start taking the analysis values.